requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
borb==2.1.17
boto3==1.28.73
botocore==1.31.73
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.4
et-xmlfile==1.1.0
fastapi==0.103.1
fonttools==4.43.0
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
lxml==4.9.3
multidict==6.0.4
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
openpyxl==3.1.2
pandas==2.1.1
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
PyPDF2==3.0.1
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
yarl==1.9.2
